backend-semantra/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
aiofiles==23.2.1
celery==5.3.4
redis==5.0.1
openai==1.3.7
google-generativeai==0.3.2
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
aiohttp==3.9.1
asyncio-throttle==1.0.2
python-slugify==8.0.1
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pandas==2.1.4
numpy==1.25.2
scikit-learn==1.3.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
pytest==7.4.3
pytest-asyncio==0.21.1 